pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crosscompute-validation"
version = "1.0.0"
dependencies = [
    "aiofiles",
    "ruamel.yaml",
]
requires-python = ">= 3.10"
authors = [
    {name = "Roy Hyunjin Han", email = "rhh@crosscompute.com"},
]
maintainers = []
description = "Configuration validation for CrossCompute tools"
readme = "README.md"
keywords = []
classifiers = [
    "Development Status :: 3 - Alpha",
    # Framework :: CrossCompute
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[project.optional-dependencies]
development = ["pre-commit"]
test = ["aiofiles", "pytest-asyncio", "pytest-cov", "pytest-xdist"]

[project.urls]
# Homepage = ""
# Documentation = ""
Repository = "https://github.com/crosscompute/crosscompute-validation"
Issues = "https://github.com/crosscompute/crosscompute-validation/issues"
# Changelog = ""

[tool.hatch.build]
packages = ["sources/crosscompute_validation"]
